refactor(SAL-72): 서버 계약을 api 계층으로 이동하고 방향 라벨 정책 추가
