+++ frontend/src/pages/verdict-board/api/verdictBoard.types.ts
